Feat: Move path completion tests to new path_test.exs
